feat: Implement restore function
